* added support for OTP API and integration on both client and server * implemented charger list page * fixed map issues with showing sites * fixed bug with stripe payment authorization

+++ composeApp/src/androidMain/kotlin/com/driivz/example/map/ManySitesClusterItem.kt
@@ -10,7 +10,7 @@ class ManySitesClusterItem(
 ) : ClusterItem {
 
     val itemPosition: LatLng = LatLng(site.latitude, site.longitude)
-    var id: Long = site.id.toLong()
+    var id: Long = site.id
 
     override fun getZIndex(): Float = 100f
     override fun equals(other: Any?): Boolean {
